- Changed preference weights from integers to floats in backend models and the LLM ranking service for finer granularity.
- Added a new 'clicked on og post' interaction type and updated existing interaction weights to float values.
- Plumbed the user object into the `EventCard` and `EventModal` components to ensure accurate user attribution for interactions.
- Implemented 'dismissed' interaction logging when a user closes the event modal.
- Refactored `logClick` payloads in the frontend to correctly structure data for the `recordInteraction` service.

+++ backend/src/models/mod.rs
@@ -176,7 +176,7 @@ pub struct UserPreference {
     pub id: Uuid,
     pub user_id: Uuid,
     pub category: String,
-    pub weight: i32,
+    pub weight: f64,
     pub created_at: DateTime<Utc>,
 }
 
@@ -184,7 +184,7 @@ pub struct UserPreference {
 #[derive(Debug, Deserialize)]
 pub struct CreateUserPreference {
     pub category: String,
-    pub weight: i32,
+    pub weight: f64,
 }
 
 // =============================================================================
@@ -249,4 +249,3 @@ pub struct UserInteractionWithEvent {
 // =============================================================================
 // SEARCH MODELS
 // =============================================================================
-
